Fix build on MacOS
The decklink routines related to strings need to be handled
differently based on the platform.  We did this back in the
beginning for capture.cpp, but the addition of the transmitter
didn't reflect those same changes.

Move the portability routines into a common header file, and
tweak the code in Config.cpp and transmitter.cpp to use those
routines.
